ARN-2702 Changes Drug History page contents if the last time any misused drug was used was >6 months ago (#828)
