+++ app/form/v1_0/steps/drug-use.ts
@@ -13,6 +13,8 @@ const stepUrls = {
   drugDetailsMoreThanSix: 'drug-details-more-than-six-months',
   drugDetailsMoreThanSixInjected: 'drug-details-more-than-six-months-injected',
   drugUseHistory: 'drug-use-history',
+  drugUseHistoryAllMoreThanSix: 'drug-use-history-more-than-six-months',
+
   summary: 'drug-use-summary',
   analysis: 'drug-use-analysis',
 }
@@ -76,7 +78,12 @@ const sectionConfig: SectionConfig = {
         drugsUseFields.sectionComplete(),
       ].flat(),
       template: templates.drugUsageNew,
-      next: stepUrls.drugUseHistory,
+      next: [
+        drugsUseFields.addDrugs.drugLastUsedFields.map(lastUsedField =>
+          nextWhen(lastUsedField, 'LAST_SIX', stepUrls.drugUseHistory),
+        ),
+        stepUrls.drugUseHistoryAllMoreThanSix,
+      ].flat(),
       sectionProgressRules: [setFieldToIncomplete(section.sectionCompleteField)],
     },
     {
@@ -92,7 +99,12 @@ const sectionConfig: SectionConfig = {
         drugsUseFields.sectionComplete(),
       ].flat(),
       template: templates.drugUsageNew,
-      next: stepUrls.drugUseHistory,
+      next: [
+        drugsUseFields.addDrugs.drugLastUsedFields.map(lastUsedField =>
+          nextWhen(lastUsedField, 'LAST_SIX', stepUrls.drugUseHistory),
+        ),
+        stepUrls.drugUseHistoryAllMoreThanSix,
+      ].flat(),
       sectionProgressRules: [setFieldToIncomplete(section.sectionCompleteField)],
     },
     {
@@ -107,7 +119,12 @@ const sectionConfig: SectionConfig = {
         drugsUseFields.sectionComplete(),
       ].flat(),
       template: templates.drugUsageNew,
-      next: stepUrls.drugUseHistory,
+      next: [
+        drugsUseFields.addDrugs.drugLastUsedFields.map(lastUsedField =>
+          nextWhen(lastUsedField, 'LAST_SIX', stepUrls.drugUseHistory),
+        ),
+        stepUrls.drugUseHistoryAllMoreThanSix,
+      ].flat(),
       sectionProgressRules: [setFieldToIncomplete(section.sectionCompleteField)],
     },
     {
@@ -124,7 +141,12 @@ const sectionConfig: SectionConfig = {
         drugsUseFields.sectionComplete(),
       ].flat(),
       template: templates.drugUsageNew,
-      next: stepUrls.drugUseHistory,
+      next: [
+        drugsUseFields.addDrugs.drugLastUsedFields.map(lastUsedField =>
+          nextWhen(lastUsedField, 'LAST_SIX', stepUrls.drugUseHistory),
+        ),
+        stepUrls.drugUseHistoryAllMoreThanSix,
+      ].flat(),
       sectionProgressRules: [setFieldToIncomplete(section.sectionCompleteField)],
     },
     {
@@ -136,7 +158,6 @@ const sectionConfig: SectionConfig = {
         drugsUseFields.drugUseHistory.drugsAffectedTheirLife,
         drugsUseFields.drugUseHistory.drugsAffectedTheirLifeDetails,
         drugsUseFields.drugUseHistory.drugsAnythingHelpedStopOrReduceUse,
-        drugsUseFields.drugUseHistory.drugsWhatCouldHelpNotUseDrugsInFuture,
         drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUse,
         drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUseDetailsPositiveChanges,
         drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUseDetailsMakingChanges,
@@ -146,6 +167,30 @@ const sectionConfig: SectionConfig = {
         drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUseDetailsDoNotWantChanges,
         drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUseDetailsDoNotWantToAnswer,
         drugsUseFields.isUserSubmitted(stepUrls.drugUseHistory),
+        drugsUseFields.sectionComplete(),
+      ].flat(),
+      next: stepUrls.summary,
+      sectionProgressRules: [setFieldToIncomplete(section.sectionCompleteField)],
+    },
+    {
+      url: stepUrls.drugUseHistoryAllMoreThanSix,
+      pageHeading: sectionHeading,
+      fields: [
+        drugsUseFields.drugUseHistory.drugsReasonsForUse,
+        drugsUseFields.drugUseHistory.drugsReasonsForUseDetails,
+        drugsUseFields.drugUseHistory.drugsAffectedTheirLife,
+        drugsUseFields.drugUseHistory.drugsAffectedTheirLifeDetails,
+        drugsUseFields.drugUseHistory.drugsAnythingHelpedStopOrReduceUse,
+        drugsUseFields.drugUseHistory.drugsWhatCouldHelpNotUseDrugsInFuture, // This one only displays if any drugs were Used more than 6 months ago
+        drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUse,
+        drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUseDetailsPositiveChanges,
+        drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUseDetailsMakingChanges,
+        drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUseDetailsWantToMakeChanges,
+        drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUseDetailsWantToMakeChangesHelp,
+        drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUseDetailsThinkingAboutChanges,
+        drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUseDetailsDoNotWantChanges,
+        drugsUseFields.drugUseHistory.drugsWantToMakeChangesToDrugUseDetailsDoNotWantToAnswer,
+        drugsUseFields.isUserSubmitted(stepUrls.drugUseHistoryAllMoreThanSix),
         drugsUseFields.sectionComplete(),
       ].flat(),
       next: stepUrls.summary,
