+++ server/views/forms/drug-usage-new.njk
@@ -53,7 +53,6 @@
 
 {% set showMoreThanSixMonths = false %}
 {% set moreThanSixMonths %}
-  <div>
     <h3 class="govuk-heading-m">Not used in the last 6 months</h3>
     {% set drugsNotUsedList = [] %}
 
@@ -73,7 +72,6 @@
     }) }}
 
     {{ renderQuestion(options.fields["not_used_in_last_six_months_details"], errors) }}
-  </div>
 {% endset %}
 
 {% block formContent %}
